+++ Code/Api/src/platform/vl53l0x_platform.c
@@ -101,7 +101,7 @@ VL53L0X_Error VL53L0X_UnlockSequenceAccess(VL53L0X_DEV Dev){
 VL53L0X_Error VL53L0X_WriteMulti(VL53L0X_DEV Dev, uint8_t index, uint8_t *pdata, uint32_t count){
 
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int = 0;
+    int8_t status_int = 0;
 	uint8_t deviceAddress;
 
     if (count>=VL53L0X_MAX_I2C_XFER_SIZE){
@@ -122,7 +122,7 @@ VL53L0X_Error VL53L0X_WriteMulti(VL53L0X_DEV Dev, uint8_t index, uint8_t *pdata,
 VL53L0X_Error VL53L0X_ReadMulti(VL53L0X_DEV Dev, uint8_t index, uint8_t *pdata, uint32_t count){
     VL53L0X_I2C_USER_VAR
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
 	uint8_t deviceAddress;
 
     if (count>=VL53L0X_MAX_I2C_XFER_SIZE){
@@ -142,7 +142,7 @@ VL53L0X_Error VL53L0X_ReadMulti(VL53L0X_DEV Dev, uint8_t index, uint8_t *pdata, 
 
 VL53L0X_Error VL53L0X_WrByte(VL53L0X_DEV Dev, uint8_t index, uint8_t data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
 	uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
@@ -157,7 +157,7 @@ VL53L0X_Error VL53L0X_WrByte(VL53L0X_DEV Dev, uint8_t index, uint8_t data){
 
 VL53L0X_Error VL53L0X_WrWord(VL53L0X_DEV Dev, uint8_t index, uint16_t data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
 	uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
@@ -172,7 +172,7 @@ VL53L0X_Error VL53L0X_WrWord(VL53L0X_DEV Dev, uint8_t index, uint16_t data){
 
 VL53L0X_Error VL53L0X_WrDWord(VL53L0X_DEV Dev, uint8_t index, uint32_t data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
 	uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
@@ -187,7 +187,7 @@ VL53L0X_Error VL53L0X_WrDWord(VL53L0X_DEV Dev, uint8_t index, uint32_t data){
 
 VL53L0X_Error VL53L0X_UpdateByte(VL53L0X_DEV Dev, uint8_t index, uint8_t AndData, uint8_t OrData){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
     uint8_t deviceAddress;
     uint8_t data;
 
@@ -211,7 +211,7 @@ VL53L0X_Error VL53L0X_UpdateByte(VL53L0X_DEV Dev, uint8_t index, uint8_t AndData
 
 VL53L0X_Error VL53L0X_RdByte(VL53L0X_DEV Dev, uint8_t index, uint8_t *data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
     uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
@@ -226,7 +226,7 @@ VL53L0X_Error VL53L0X_RdByte(VL53L0X_DEV Dev, uint8_t index, uint8_t *data){
 
 VL53L0X_Error VL53L0X_RdWord(VL53L0X_DEV Dev, uint8_t index, uint16_t *data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
     uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
@@ -241,7 +241,7 @@ VL53L0X_Error VL53L0X_RdWord(VL53L0X_DEV Dev, uint8_t index, uint16_t *data){
 
 VL53L0X_Error  VL53L0X_RdDWord(VL53L0X_DEV Dev, uint8_t index, uint32_t *data){
     VL53L0X_Error Status = VL53L0X_ERROR_NONE;
-    int32_t status_int;
+    int8_t status_int;
     uint8_t deviceAddress;
 
     deviceAddress = Dev->I2cDevAddr;
